以委托的方式使用 SharedPreferences

+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/DetailActivity.kt
@@ -35,7 +35,7 @@ class DetailActivity : AppCompatActivity(), ToolbarManager {
         enableHomeAsUp { onBackPressed() }
 
         doAsync {
-            val result = RequestDayForecastCommand(intent.getLongExtra(ID, -1)).execute()
+            val result = RequestDayForecastCommand(intent.getIntExtra(ID, -1)).execute()
             uiThread { bindForecast(result) }
         }
     }
