+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/data/db/DbClasses.kt
@@ -10,11 +10,11 @@ import kotlin.collections.HashMap
  * 通过 map 的使用，我们可以用很简单的方式把类转换为数据表，反之依然。
  */
 class CityForecast(val map: MutableMap<String, Any?>, val dailyForecast: List<DayForecast>) {
-    var _id: Long by map
+    var _id: Int by map
     var city: String by map
     var country: String by map
 
-    constructor(id: Long, city: String, country: String, dailyForecast: List<DayForecast>) : this(HashMap(), dailyForecast) {
+    constructor(id: Int, city: String, country: String, dailyForecast: List<DayForecast>) : this(HashMap(), dailyForecast) {
         this._id = id
         this.city = city
         this.country = country
@@ -23,15 +23,15 @@ class CityForecast(val map: MutableMap<String, Any?>, val dailyForecast: List<Da
 
 
 class DayForecast(var map: MutableMap<String, Any?>) {
-    var _id: Long by map
+    var _id: Int by map
     var date: Long by map
     var description: String by map
     var high: Int by map
     var low: Int by map
     var iconUrl: String by map
-    var cityId: Long by map
+    var cityId: Int by map
 
-    constructor(date: Long, description: String, high: Int, low: Int, iconUrl: String, cityId: Long) : this(HashMap()) {
+    constructor(date: Long, description: String, high: Int, low: Int, iconUrl: String, cityId: Int) : this(HashMap()) {
         this.date = date
         this.description = description
         this.high = high
