+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/data/server/ServerDataMapper.kt
@@ -12,7 +12,7 @@ import com.badmask_zly.kotlinlearnmiddle.domain.model.Forecast2 as ModelForecast
  */
 class ServerDataMapper {
 
-    fun convertToDomain(zipCode: Long, forecast: ForecastResult) = with(forecast) {
+    fun convertToDomain(zipCode: Int, forecast: ForecastResult) = with(forecast) {
         ForecastList2(zipCode, city.name, city.country, convertForecastListToDomain(list))
     }
