+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/domain/commands/RequestDayForecastCommand.kt
@@ -6,6 +6,6 @@ import com.badmask_zly.kotlinlearnmiddle.domain.model.Forecast2
 /**
  * Created by badmask_zly on 2017/6/22.
  */
-class RequestDayForecastCommand(val id: Long, val forecastProvider: ForecastProvider = ForecastProvider()) : Command<Forecast2> {
+class RequestDayForecastCommand(val id: Int, val forecastProvider: ForecastProvider = ForecastProvider()) : Command<Forecast2> {
     override fun execute() = forecastProvider.requestForecast(id)
 }
