+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/domain/commands/RequestForecastCommand2.kt
@@ -8,7 +8,7 @@ import com.badmask_zly.kotlinlearnmiddle.domain.model.ForecastList2
  * 用新的写法替换原来的 RequestForecastCommand
  * 相较与原来的内容，RequestForecastCommand2 不会再直接与服务端交互，也不会转换数据到 doamin model
  */
-class RequestForecastCommand2(val zipCode: Long,
+class RequestForecastCommand2(val zipCode: Int,
                               val forecastProvider: ForecastProvider = ForecastProvider()) : Command<ForecastList2> {
 
     companion object {
