+++ app/src/main/java/com/badmask_zly/kotlinlearnmiddle/domain/model/DomainClasses.kt
@@ -1,6 +1,6 @@
 package com.badmask_zly.kotlinlearnmiddle.domain.model
 
-data class ForecastList2(val id: Long, val city: String, val country: String,
+data class ForecastList2(val id: Int, val city: String, val country: String,
         val dailyForecast: List<Forecast2>) {
 
     val size: Int
@@ -9,6 +9,6 @@ data class ForecastList2(val id: Long, val city: String, val country: String,
     operator fun get(position: Int) = dailyForecast[position]
 }
 
-data class Forecast2(val id: Long, val date: Long, val description: String, val high: Int,
+data class Forecast2(val id: Int, val date: Long, val description: String, val high: Int,
         val low: Int,
                     val iconUrl: String)
